kokkos-kernels: fix singleton multiple definitions
Resolve issues with Magma and Cuda singleton multiple definitions
Reported in kokkos/kokkos-kernels#2175
Patch generated from kokkos/kokkos-kernels#2176
