+++ packages/kokkos-kernels/blas/tpls/KokkosBlas_Cuda_tpl.cpp
@@ -16,3 +16,4 @@
 #include <Kokkos_Core.hpp>
 #include <KokkosKernels_config.h>
 #include <KokkosBlas_Cuda_tpl.hpp>
+#include <KokkosBlas_Magma_tpl.hpp>
