+++ packages/kokkos-kernels/blas/tpls/KokkosBlas_Magma_tpl.hpp
@@ -13,30 +13,29 @@
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 //
 //@HEADER
-#ifndef KOKKOSBLAS_CUDA_TPL_HPP_
-#define KOKKOSBLAS_CUDA_TPL_HPP_
+#ifndef KOKKOSBLAS_MAGMA_TPL_HPP_
+#define KOKKOSBLAS_MAGMA_TPL_HPP_
 
-#if defined(KOKKOSKERNELS_ENABLE_TPL_CUBLAS)
-#include <KokkosBlas_tpl_spec.hpp>
+#if defined(KOKKOSKERNELS_ENABLE_TPL_MAGMA)
+#include <KokkosBlas_magma.hpp>
 
 namespace KokkosBlas {
 namespace Impl {
 
-CudaBlasSingleton::CudaBlasSingleton() {
-  cublasStatus_t stat = cublasCreate(&handle);
-  if (stat != CUBLAS_STATUS_SUCCESS)
-    Kokkos::abort("CUBLAS initialization failed\n");
+MagmaSingleton::MagmaSingleton() {
+  magma_int_t stat = magma_init();
+  if (stat != MAGMA_SUCCESS) Kokkos::abort("MAGMA initialization failed\n");
 
-  Kokkos::push_finalize_hook([&]() { cublasDestroy(handle); });
+  Kokkos::push_finalize_hook([&]() { magma_finalize(); });
 }
 
-CudaBlasSingleton& CudaBlasSingleton::singleton() {
-  static CudaBlasSingleton s;
+MagmaSingleton& MagmaSingleton::singleton() {
+  static MagmaSingleton s;
   return s;
 }
 
 }  // namespace Impl
 }  // namespace KokkosBlas
-#endif  // defined (KOKKOSKERNELS_ENABLE_TPL_CUBLAS)
+#endif  // defined(KOKKOSKERNELS_ENABLE_TPL_MAGMA)
 
-#endif  // KOKKOSBLAS_CUDA_TPL_HPP_
+#endif  // KOKKOSBLAS_MAGMA_TPL_HPP_
