+++ packages/kokkos-kernels/lapack/tpls/KokkosLapack_Magma_tpl.hpp
@@ -13,30 +13,29 @@
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 //
 //@HEADER
-#ifndef KOKKOSLAPACK_CUDA_TPL_HPP_
-#define KOKKOSLAPACK_CUDA_TPL_HPP_
+#ifndef KOKKOSLAPACK_MAGMA_TPL_HPP_
+#define KOKKOSLAPACK_MAGMA_TPL_HPP_
 
-#if defined(KOKKOSKERNELS_ENABLE_TPL_CUSOLVER)
-#include "KokkosLapack_cusolver.hpp"
+#if defined(KOKKOSKERNELS_ENABLE_TPL_MAGMA)
+#include <KokkosLapack_magma.hpp>
 
 namespace KokkosLapack {
 namespace Impl {
 
-CudaLapackSingleton::CudaLapackSingleton() {
-  cusolverStatus_t stat = cusolverDnCreate(&handle);
-  if (stat != CUSOLVER_STATUS_SUCCESS)
-    Kokkos::abort("CUSOLVER initialization failed\n");
+MagmaSingleton::MagmaSingleton() {
+  magma_int_t stat = magma_init();
+  if (stat != MAGMA_SUCCESS) Kokkos::abort("MAGMA initialization failed\n");
 
-  Kokkos::push_finalize_hook([&]() { cusolverDnDestroy(handle); });
+  Kokkos::push_finalize_hook([&]() { magma_finalize(); });
 }
 
-CudaLapackSingleton& CudaLapackSingleton::singleton() {
-  static CudaLapackSingleton s;
+MagmaSingleton& MagmaSingleton::singleton() {
+  static MagmaSingleton s;
   return s;
 }
 
 }  // namespace Impl
 }  // namespace KokkosLapack
-#endif  // defined (KOKKOSKERNELS_ENABLE_TPL_CUSOLVER)
+#endif  // defined(KOKKOSKERNELS_ENABLE_TPL_MAGMA)
 
-#endif  // KOKKOSLAPACK_CUDA_TPL_HPP_
+#endif  // KOKKOSLAPACK_MAGMA_TPL_HPP_
